swapped contents of test main and example main
